Finish tidying nature spirit
