+++ data/quest/members/nature_spirit/nature_spirit.areas.toml
@@ -1,3 +1,7 @@
 [grotto_island]
 x = [3438, 3444]
 y = [9735, 9742]
+
+[grotto]
+x = [3438, 3444]
+y = [9735, 9742]
